+++ Util.h
@@ -19,7 +19,7 @@ const int Max_Ephemral_Public_Key_Size = 2048;
 const int Encrypted_Signature_Size = 272; // 256 for the signature +16 for the aes padding block
 const int CBC_IV_Length = EVP_CIPHER_iv_length(EVP_aes_128_cbc());
 
-bool receiveEphemralPublicKey(int clientSocket, EVP_PKEY *&deserializedKey, unsigned char *&serializedKey, int &serializedKeyLength);
+bool receiveEphemralPublicKey(int clientSocket, EVP_PKEY *&deserializedKey, unsigned char *&serializedKey, size_t &serializedKeyLength);
 bool generateDigitalSignature(unsigned char *data, size_t dataLength, EVP_PKEY *privateKey, unsigned char *&signature, unsigned int &signatureLength);
 bool verifyDigitalSignature(unsigned char *data, size_t dataLength, unsigned char *signature, unsigned int signatureLength, EVP_PKEY *publicKey);
 bool computeSHA256Digest(unsigned char *data, size_t dataLength, unsigned char *&digest, unsigned int &digestLength);
